Always use the protocol that the Sandstorm instance is served at
While I originally used protocol-relative links, this caused issues if the
Sandstorm server and the page the tracker was embedded in were of different
HTTPS-ness.  If you try to hit Sandstorm on HTTPS when it's expecting HTTP,
you'll probably get no reply.  If you try to hit Sandstorm on HTTP when it's
expecting HTTPS, you'll probably get a redirect from the HTTPS frontend, and
then the browser will send a GET without the Authorization: header, so the
request won't reach the backend.  In both cases, you'd get silent failures,
and it'd look like Sandstorm's fault.

Now, if you mix HTTP-site and HTTPS-sandstorm, things will just work, and if
you mix HTTPS-site with HTTP-sandstorm, you'll get the mixed content
warning/failure that you deserve, which is much better.